+++ rll/metadata.rb
@@ -4,14 +4,12 @@ license     "see LICENSE file in repository root"
 description "Base scripts for RightLink10 on Linux (RLL) to initialize basic functionality"
 version     '10.1.5'
 
-recipe      "rll::wait-for-eip", "Wait for external IP address to be assigned (EC2 issue)"
-recipe      "rll::security_updates", "Installs security updates"
-recipe      "rll::setup_hostname", "Changes the hostname of the server"
 recipe      "rll::collectd", "Installs and configures collectd for RightScale monitoring"
-recipe      "rll::upgrade", "Check whether a RightLink upgrade is available and do the upgrade"
+recipe      "rll::setup-automatic-upgrade", "Periodically checks if an upgrade is available and upgrade if there is."
+recipe      "rll::setup-hostname", "Changes the hostname of the server"
 recipe      "rll::shutdown-reason", "Print out the reason for shutdown"
-recipe      "rll::setup_automatic_upgrade", "Periodically checks if an upgrade is available and upgrade if there is."
-recipe      "rll::test-script", "Test operational script, used by righlinklite/tester"
+recipe      "rll::upgrade", "Check whether a RightLink upgrade is available and do the upgrade"
+recipe      "rll::wait-for-eip", "Wait for external IP address to be assigned (EC2 issue)"
 
 attribute   "SERVER_HOSTNAME",
   :display_name => "Hostname for this server",
@@ -22,7 +20,7 @@ attribute   "SERVER_HOSTNAME",
   :required => "optional",
   :type => "string",
   :default => "",
-  :recipes => ["rll::setup_hostname"]
+  :recipes => ["rll::setup-hostname"]
 
 attribute   "COLLECTD_SERVER",
   :display_name => "RightScale monitoring server to send data to",
@@ -44,7 +42,7 @@ attribute   "ENABLE_AUTO_UPGRADE",
   :type => "string",
   :default => "true",
   :choice => ["true", "false"],
-  :recipes => ["rll::setup_automatic_upgrade"]
+  :recipes => ["rll::setup-automatic-upgrade"]
 
 attribute   "UPGRADES_FILE_LOCATION",
   :display_name => "External location of 'upgrades' file",
@@ -52,41 +50,3 @@ attribute   "UPGRADES_FILE_LOCATION",
   :type => "string",
   :default => "https://rightlink.rightscale.com/rightlink/upgrades",
   :recipes => ["rll::upgrade"]
-
-#####################
-# Inputs for internal testing below
-#####################
-
-attribute   "UTF8_STRING_INPUT",
-  :display_name => "UTF8 String Input",
-  :required => "recommended",
-  :type => "string",
-  :default => "hello,здравствуйте",
-  :recipes => ["rll::test-script"]
-
-attribute   "STRING_INPUT_1",
-  :display_name => "String Input 1",
-  :required => "recommended",
-  :type => "string",
-  :default => "Default String Input 1",
-  :recipes => ["rll::test-script"]
-
-attribute   "STRING_INPUT_2",
-  :display_name => "String Input 2",
-  :required => "recommended",
-  :type => "string",
-  :default => "Default String Input 2",
-  :recipes => ["rll::test-script"]
-
-attribute   "CRED_INPUT",
-  :display_name => "Credential Input",
-  :required => "recommended",
-  :type => "string",
-  :recipes => ["rll::test-script"]
-
-attribute   "ARRAY_INPUT_1",
-  :display_name => "Array Input 1",
-  :required => "recommended",
-  :type => "array",
-  :default => ["ARRAY_ITEM_1", "ARRAY_ITEM_2", "ARRAY_ITEM_3"],
-  :recipes => ["rll::test-script"]