+++ rll/setup-hostname.sh
@@ -1,5 +1,21 @@
 #! /bin/bash -e
 
+# ---
+# RightScript Name: RL10 Linux Setup Hostname
+# Description: Changes the hostname of the server.
+# Inputs:
+#   SERVER_HOSTNAME:
+#     Input Type: single
+#     Category: RightScale
+#     Description: The server's hostname is set to the longest valid prefix or suffix of this variable.
+#       E.g. 'my.example.com V2', 'NEW my.example.com', and 'database.io my.example.com' all set the hostname to 'my.example.com'.
+#       Set to an empty string to avoid any change to the hostname.
+#     Default: text:
+#     Required: false
+#     Advanced: true
+# ...
+#
+
 # The server's hostname is set to the longest valid prefix or suffix of
 # this SERVER_HOSTNAME variable eg 'my.example.com V2', 'NEW my.example.com', and
 # 'database.io my.example.com' all set the hostname to 'my.example.com'.