+++ CHANGES.md
@@ -5,6 +5,9 @@ Changelog for the RightLink10 Base ServerTemplate
 ------
 - Fix collectd.sh script failure on CentOS 6 when SELinux is enabled.
 - Update rlw::shutdown-reason to have parity with Linux counterpart.
+- Move example scripts to their own cookbooks
+- Add RightScript metadata to scripts
+- Remove testing scripts
 
 10.1.4
 ------